package/libudev: make it selectable
Usually, it is not possible for a package to select a virtual package it
depends on, and expects a provider to suddenly be available.

However, for libudev, this is slightly simpler: there will be only two
providers ever: eudev and systemd. Both are incompatible one with the
other, systemd is only available if chosen as an init system.

So, if systemd is not enabled (as an init system), then we can
forcefully enable eudev as the provider for libudev.

However, this can cause circular dependencies in Kconfig, du to a rather
complex chain of select and depends. For example, when libinput is
converting in a later patch, we'd end up with:

    package/weston/Config.in:6:error: recursive dependency detected!
    package/weston/Config.in:6:     symbol BR2_PACKAGE_WESTON depends on BR2_PACKAGE_HAS_UDEV
    package/udev/Config.in:1:       symbol BR2_PACKAGE_HAS_UDEV is selected by BR2_PACKAGE_EUDEV_DAEMON
    package/eudev/Config.in:24:     symbol BR2_PACKAGE_EUDEV_DAEMON depends on BR2_PACKAGE_EUDEV
    package/eudev/Config.in:1:      symbol BR2_PACKAGE_EUDEV is selected by BR2_PACKAGE_LIBUDEV
    package/libudev/Config.in:1:    symbol BR2_PACKAGE_LIBUDEV is selected by BR2_PACKAGE_LIBINPUT
    package/libinput/Config.in:1:   symbol BR2_PACKAGE_LIBINPUT is selected by BR2_PACKAGE_WESTON

The root cause being a mis-match of select and depends on in the same
chain, where Kconfig can not determine that an external condition will
ensure that the chain is in fact broken at some point.

To fix that, we move the symbol for the eudev daemon out of the eudev
if-block. This breaks the chain.

Yet, we still want to tell the user that the daemon is enabled, so we
introduce a dummy symbol, just for the sake of having a forced-on
prompt.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>