package/kodi: bump to 22.0b1-Piers
Removed patch from kodi-texturepacker which is included in this release.

Removed patches 0002-0005 which are included in this release.

Added dependency to exiv2:
https://github.com/xbmc/xbmc/commit/197a3e6d2e80878f5f3f827b3a9073118b59bd51

Replaced dependency rapidjson with nlohmann-json:
https://github.com/xbmc/xbmc/commit/d6d2cac380ea127c2a64a998a31622debe149d3b

The mandatory dependency to libcdio was moved to optical drives support:
https://github.com/xbmc/xbmc/commit/11289a597a88d809f39d3c23aa7eae6325a5263c

Remove -DCMAKE_MODULE_PATH from host-kodi-texturepacker, which was used
to find FindLzo2.cmake, after upstream commit
https://github.com/xbmc/xbmc/commit/3eb5196bf79759682aa0eb20b0b4830c083f81a5
added a version of FindLzo2.cmake specifically for the host build of
texturepacker.

Bumped versions of java tarballs following upstream commits
https://github.com/xbmc/xbmc/commit/aee5f9cd76ac13a9b3089e0b3458163134891312
https://github.com/xbmc/xbmc/commit/82a339d1ada00b8e092181003a029526bfdb0aef
https://github.com/xbmc/xbmc/commit/0c9d46c53bdc660479f6507b0b02784d8b4e824e

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>