package/{rust, rust-bin}: install rustfmt
Rust-for-Linux expects rustfmt to be available; otherwise the kernel
emits a "rustfmt not found" warning and the generated bindings are not
correct. Provide rustfmt from rust-bin by adding it to the installed
components.

Note: the dist tarball exposes the component as "rustfmt-preview"
(as listed in 'components' file), so we use that name.

Signed-off-by: El Mehdi YOUNES <elmehdi.younes@smile.fr>
[Bernd: added rustfmt-preview to rust package]
Signed-off-by: Bernd Kuhls <bernd@kuhls.net>