package/dvbsnoop: remove package
The package is causing build errors with newer gcc versions:

sectables.c:331:22: error: initialization of 'void (*)(void)' from
 incompatible pointer type 'void (*)(unsigned char *, int)'
 [-Wincompatible-pointer-types]
  331 |      {  0x00, 0x00,  section_PAT        },

and many more.

The latest release dates back to 2007, the latest change in the CVS-
based source repository[1] (downloaded as dvbsnoop.zip) is also from
2007. No newer or maintained forks were found.

State of upstream ticket system[2]:

  Bugs: last update 2012
  Support Requests: No open tickets found.
  Patches: last update 2010
  Feature Requests: 1 from 2021, the rest 2012 and older

Instead of fixing build errors for an unmaintained project which is not
used by any other package we remove it from buildroot.

Fixes:
https://autobuild.buildroot.net/results/ea6/ea6786248b9ecf0eb1c1c1f7cdd0202462b75e00/

[1] https://sourceforge.net/p/dvbsnoop/cvs/
[2] https://sourceforge.net/p/dvbsnoop/_list/tickets